+++ src/CellData/SurfaceDiffOps.jl
@@ -113,34 +113,35 @@ function Δs(f::Function,
     _compose(Δs_parametric_space, InvAmbientMapCellField(Ω_atlas))
 end
 
+# returns the contravariant component of surface gradient
 function _∇s_no_ad(f, Ω_atlas)
   # sgrad(f::Function) = m -> sgrad(f,m)
-  # sgrad(f::Function,m::Field) = αβ -> J(m,αβ) ⋅
-  #                                     (inv_metric(m,αβ) ⋅ gradient(f(m))(αβ) )
+  # sgrad(f::Function,m::Field) = αβ -> (inv_metric(m,αβ) ⋅ gradient(f(m))(αβ) )
   ambient_map_cf = AmbientMapCellField(Ω_atlas)
   inv_metric_cf = InvMetricCellField(Ω_atlas)
   covariant_basis_cf = transpose∘∇(ambient_map_cf)
   gradient_f_cf = (∇(f)∘ambient_map_cf)⋅covariant_basis_cf
-  covariant_basis_cf⋅(inv_metric_cf⋅gradient_f_cf)
+  (inv_metric_cf⋅gradient_f_cf)
 end
 
+# returns the contravariant component of surface gradient
 function _∇s_ad(f, Ω_atlas)
   # sgrad(f::Function) = m -> sgrad(f,m)
-  # sgrad(f::Function,m::Field) = αβ -> J(m,αβ) ⋅
-  #                                     (inv_metric(m,αβ) ⋅ gradient(f(m))(αβ) )
+  # sgrad(f::Function,m::Field) = αβ -> (inv_metric(m,αβ) ⋅ gradient(f(m))(αβ) )
   ambient_map_cf = AmbientMapCellField(Ω_atlas)
   ambient_maps = Gridap.CellData.get_data(ambient_map_cf)
   cell_field = lazy_map(m->GenericField(sgrad(_fm(f,m),m)),ambient_maps)
   CellData.GenericCellField(cell_field,Ω_atlas,PhysicalDomain())
 end
 
-
+# returns the contravariant component of surface gradient
 function ∇s(f::Function,
             Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
             use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
   use_automatic_differentiation ? _∇s_ad(f, Ω_atlas) : _∇s_no_ad(f, Ω_atlas)
 end
 
+# returns the contravariant component of surface gradient
 function ∇s(f::Function,
             Ω_atlas::AdaptedTriangulation{Dc,Da,<:BFTATDMIM{Dc,Dc,Da,G,A,P,C,O}};
             use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
@@ -151,25 +152,34 @@ end
 function ∇s(f::Function,
             Ω_atlas::BFTATDMEM{Dc,Dc,Da,G,A,P,C,O};
             use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
-  ∇s_parametric_space = use_automatic_differentiation ? _∇s_ad(f, Ω_atlas) : _∇s_no_ad(f, Ω_atlas)
+  ∇s_parametric_space_contra = use_automatic_differentiation ? _∇s_ad(f, Ω_atlas) : _∇s_no_ad(f, Ω_atlas)
+  ambient_map_cf = AmbientMapCellField(Ω_atlas)
+  covariant_basis_cf = transpose∘∇(ambient_map_cf)
+  ∇s_parametric_space = covariant_basis_cf ⋅ ∇s_parametric_space_contra
   _compose(∇s_parametric_space, InvAmbientMapCellField(Ω_atlas))
 end
 
 function ∇s(f::Function,
             Ω_atlas::AdaptedTriangulation{Dc,Da,<:BFTATDMEM{Dc,Dc,Da,G,A,P,C,O}};
             use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
-  ∇s_parametric_space = use_automatic_differentiation ? _∇s_ad(f, Ω_atlas) : _∇s_no_ad(f, Ω_atlas)
+  ∇s_parametric_space_contra = use_automatic_differentiation ? _∇s_ad(f, Ω_atlas) : _∇s_no_ad(f, Ω_atlas)
+  ambient_map_cf = AmbientMapCellField(Ω_atlas)
+  covariant_basis_cf = transpose∘∇(ambient_map_cf)
+  ∇s_parametric_space = covariant_basis_cf ⋅ ∇s_parametric_space_contra
   _compose(∇s_parametric_space, InvAmbientMapCellField(Ω_atlas))
 end
 
+# return the contravariant componet
 function _skew_∇s_no_ad(f, Ω_atlas)
   ambient_map_cf = AmbientMapCellField(Ω_atlas)
   meas_cf = MeasureCellField(Ω_atlas)
   J_cf = transpose∘∇(ambient_map_cf)
   grad_f_cf = (∇(f)∘ambient_map_cf)⋅J_cf
-  skew_grad_parametric = J_cf⋅(perp∘grad_f_cf)*(1.0/meas_cf) 
+  skew_grad_parametric = (perp∘grad_f_cf)*(1.0/meas_cf)
+  skew_grad_parametric
 end
 
+# return the contravariant componet
 function _skew_∇s_ad(f, Ω_atlas)
   ambient_map_cf = AmbientMapCellField(Ω_atlas)
   ambient_maps = Gridap.CellData.get_data(ambient_map_cf)
@@ -177,9 +187,18 @@ function _skew_∇s_ad(f, Ω_atlas)
   CellData.GenericCellField(cell_field,Ω_atlas,PhysicalDomain())
 end
 
+# return the contravariant componet
 function skew_∇s(f::Function, Ω_atlas::BFTATDMIM{2,2,Da,G,A,P,C,O};
                    use_automatic_differentiation=false) where {Da, G, A, P, C, O}
    use_automatic_differentiation ? _skew_∇s_ad(f, Ω_atlas) : _skew_∇s_no_ad(f, Ω_atlas)
+end
+
+# return the contravariant componet
+function skew_∇s(f::Function,
+                  Ω_atlas::AdaptedTriangulation{2,2,<:BFTATDMIM{2,2,Da,G,A,P,C,O}};
+                  use_automatic_differentiation=false) where {Da, G, A, P, C, O}
+  skew_∇s_trian = use_automatic_differentiation ? _skew_∇s_ad(f, Ω_atlas.trian) : _skew_∇s_no_ad(f, Ω_atlas.trian)
+  Gridap.CellData.GenericCellField(get_data(skew_∇s_trian), Ω_atlas, Gridap.CellData.DomainStyle(skew_∇s_trian))
 end
 
 function skew_∇s(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
@@ -208,7 +227,7 @@ end
 
 function _divs_no_ad(f, Ω_atlas)
     # 1/m * div( m * (J^†⋅(f∘ϕ)) ), where J^†=inv(g)⋅Jᵀ
-    # grad(m)⋅(J^†⋅(f∘ϕ)) + 
+    # grad(m)⋅(J^†⋅(f∘ϕ)) +
     # div((J^†⋅(f∘ϕ))) = tr(grad(J^†):(f∘ϕ)) + tr(J^†⋅grad(f∘ϕ))
     # grad(J^†) = grad(inv(g)⋅Jᵀ) = grad(inv(g))⋅Jᵀ + inv(g)⊙grad(Jᵀ)
     metric_cf = MetricCellField(Ω_atlas)
@@ -219,11 +238,11 @@ function _divs_no_ad(f, Ω_atlas)
     f_cf = f∘ambient_map_cf
     grad_f_cf = ∇(f)∘ambient_map_cf
     Jt_cf = ∇(ambient_map_cf)
-    
+
     # grad(inv(g))⋅Jᵀ
     grad_inv_metric_cf = ∇(inv_metric_cf)
     trace_1=Operation(tr)((grad_inv_metric_cf⋅Jt_cf)⋅f_cf)
-    
+
     # inv(g)⋅grad(Jᵀ)
     trace_2 = Operation(tr)((inv_metric_cf ⋅ ∇(Jt_cf))⋅f_cf)
 
@@ -233,20 +252,20 @@ function _divs_no_ad(f, Ω_atlas)
 
     grad_meas_cf = (deriv_sqrt∘det∘metric_cf)*
                    Operation(cpAB)(deriv_det∘metric_cf,gradient(metric_cf))
-   
 
-    return (1.0/meas_cf)*(meas_cf*(trace_1+trace_2+trace_3) + 
+
+    return (1.0/meas_cf)*(meas_cf*(trace_1+trace_2+trace_3) +
                            grad_meas_cf⋅(inv_metric_cf⋅Jt_cf⋅f_cf))
 end
 
 function _skew_divs_no_ad(f, Ω_atlas)
     # -1/m * div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) ), where J^†=inv(g)⋅Jᵀ
     # div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) )
-    # div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) ) = 
+    # div( m^2 * inv(g) R(J^†⋅(f∘ϕ)) ) =
     #   grad(m^2)⋅(inv(g) R(J^†⋅(f∘ϕ))) + m^2 * div(inv(g) R(J^†⋅(f∘ϕ)))
     # div(inv(g) R(J^†⋅(f∘ϕ))) = tr(grad(inv(g))⋅R(J^†⋅(f∘ϕ))) + tr(inv(g)⋅grad(R(J^†⋅(f∘ϕ))))
     #
-    Gridap.Helpers.@notimplemented "skew_divs without automatic differentiation is not implemented yet"  
+    Gridap.Helpers.@notimplemented "skew_divs without automatic differentiation is not implemented yet"
 end
 
 function _skew_divs_ad(f, Ω_atlas)
@@ -256,8 +275,8 @@ function _skew_divs_ad(f, Ω_atlas)
   CellData.GenericCellField(cell_field,Ω_atlas,PhysicalDomain())
 end
 
-# Surface divergence of an ambient vector-valued function which is 
-# pulled back using the pseudo-inverse of the jacobian of the ambient 
+# Surface divergence of an ambient vector-valued function which is
+# pulled back using the pseudo-inverse of the jacobian of the ambient
 # map without multiplying by the measure
 function divs(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
               use_automatic_differentiation=false) where {Dc, Da, G, A, P, C, O}
@@ -365,7 +384,7 @@ curls(u,m) = x-> 1.0/sqrtg(m,x)*metric(m,x)⋅curl(Jtu(u,m))(x)
 curls_curls(u, m) = x -> 1.0/sqrtg(m,x)*metric(m,x)⋅curl(curls(u,m))(x)
 
 ## surface divergence
-_divs(u,m) = x -> sqrtg(m)(x)*inv(J(m,x))⋅u(m)(x)
+_divs(u,m) = x -> sqrtg(m)(x)*contra_v(u,m)(x) # use contra_v so that pinv dispatches for 2D/3D appropriately
 divs(u, m) = x -> 1/sqrtg(m)(x)*(divergence(_divs(u,m))(x))
 
 ### covariant vector surfgrad(surfdiv u)
@@ -380,10 +399,10 @@ function _vecΔs_ad(f, Ω_atlas)
 end
 
 function _vecΔs_no_ad(f, Ω_atlas)
-  Gridap.Helpers.@notimplemented "vecΔs without automatic differentiation is not implemented yet"  
+  Gridap.Helpers.@notimplemented "vecΔs without automatic differentiation is not implemented yet"
 end
 
-# Returns the co-vector components of the vector surface laplacian applied to the 
+# Returns the co-vector components of the vector surface laplacian applied to the
 # ambient vector-valued function f
 function vecΔs(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
                 use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
@@ -415,10 +434,10 @@ function _curls_ad(f, Ω_atlas)
 end
 
 function _curls_no_ad(f, Ω_atlas)
-  Gridap.Helpers.@notimplemented "curls without automatic differentiation is not implemented yet"  
+  Gridap.Helpers.@notimplemented "curls without automatic differentiation is not implemented yet"
 end
 
-# Returns the co-vector components of the surface curl operator applied to the 
+# Returns the co-vector components of the surface curl operator applied to the
 # ambient vector-valued function f
 function curls(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
                 use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
@@ -429,4 +448,43 @@ function curls(f::Function, Ω_atlas::AdaptedTriangulation{Dc,Da,<:BFTATDMIM{Dc,
                 use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
   curls_trian = use_automatic_differentiation ? _curls_ad(f, Ω_atlas.trian) : _curls_no_ad(f, Ω_atlas.trian)
   Gridap.CellData.GenericCellField(get_data(curls_trian), Ω_atlas, Gridap.CellData.DomainStyle(curls_trian))
+end
+
+
+#### Contravariant components of 2D vector laplacian, required for Stokes
+# Contravariant components of the ∇^perp (∇^perp ⋅ ̃u)
+# = J*    1/√g R gradient ( -1/√g div(R g u)     )
+# where u is contravariant component of ̃u, Recall Rg = (√g)^2 g^{-1} R
+curly_curl(u, m) = x -> 1/sqrtg(m)(x)*perp(gradient(skew_surfdiv(contra_v(u),m))(x))
+
+# Contravariant component of ∇(∇⋅ ̃u)
+contra_grads_divs(u, m) = x-> inv_metric(m,x)⋅grads_divs(u,m)(x)
+
+# Contravariant component of the surface vector laplacian in 2D
+vec_laps_2D(u,m) = x ->  contra_grads_divs(u,m)(x) -1.0*curly_curl(u,m)(x)
+
+function _vecΔs_ad_2D(f, Ω_atlas)
+  ambient_map_cf = AmbientMapCellField(Ω_atlas)
+  ambient_maps = Gridap.CellData.get_data(ambient_map_cf)
+  cell_field = lazy_map(m->GenericField(vec_laps_2D(_fm(f,m),m)),ambient_maps)
+  CellData.GenericCellField(cell_field,Ω_atlas,PhysicalDomain())
+end
+
+function _vecΔs_no_ad_2D(f, Ω_atlas)
+  Gridap.Helpers.@notimplemented "vecΔs_2D without automatic differentiation is not implemented yet"
+end
+
+# Returns the contravariant components of the vector surface laplacian applied to the
+# ambient vector-valued function f in 2D
+function vecΔs_2D(f::Function, Ω_atlas::BFTATDMIM{Dc,Dc,Da,G,A,P,C,O};
+                use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
+  use_automatic_differentiation ? _vecΔs_ad_2D(f, Ω_atlas) : _vecΔs_no_ad_2D(f, Ω_atlas)
+end
+
+
+function vecΔs_2D(f::Function,
+              Ω_atlas::AdaptedTriangulation{Dc,Da,<:BFTATDMIM{Dc,Dc,Da,G,A,P,C,O}};
+              use_automatic_differentiation=true) where {Dc, Da, G, A, P, C, O}
+  vecΔs_2D_trian = use_automatic_differentiation ? _vecΔs_ad_2D(f, Ω_atlas.trian) : _vecΔs_no_ad_2D(f, Ω_atlas.trian)
+  Gridap.CellData.GenericCellField(get_data(vecΔs_2D_trian), Ω_atlas, Gridap.CellData.DomainStyle(vecΔs_2D_trian))
 end