+++ test/SurfaceStokes/seq/SurfaceStokesTests.jl
@@ -1,0 +1,7 @@
+include("../SurfaceStokes_TaylorHood.jl")
+
+## Serial model: 2D
+n_ref_lvls = 4
+radius = 1.0
+models = generate_refined_models(n_ref_lvls, CubedSphereMesh(radius), IntrinsicManifold())
+SurfaceStokes_TaylorHood.main(models;jumpCalcs=true)