+++ test/SurfaceStokes/seq/VectorLaplacianTests_2D.jl
@@ -1,0 +1,7 @@
+include("../VectorLaplacian2D.jl")
+
+## Serial model: 2D
+n_ref_lvls = 4
+radius = 1.0
+models = generate_refined_models(n_ref_lvls, CubedSphereMesh(radius), IntrinsicManifold())
+VectorLaplacian2D.main(models)