+++ test/SurfaceStokes/seq/runtests.jl
@@ -1,0 +1,4 @@
+using Test
+
+@testset "VectorLaplacian_2D" begin include("VectorLaplacianTests_2D.jl") end
+@testset "SurfaceStokes_2D" begin include("SurfaceStokesTests.jl") end