+++ test/runtests.jl
@@ -52,6 +52,10 @@ if TESTCASE ∈ ("all", "seq", "seq-benchmark")
   include("../benchmark/test/runtests.jl")
 end
 
+if TESTCASE ∈ ("all", "seq", "seq-stokes")
+  include("SurfaceStokes/seq/runtests.jl")
+end
+
 # MPI tests
 if TESTCASE ∈ ("all", "mpi", "mpi-atlas-discrete-models")
    include("AtlasDiscreteModels/mpi/runtests.jl")
@@ -91,4 +95,8 @@ end
 
 if TESTCASE ∈ ("all", "mpi", "mpi-ambient-model")
   include("AmbientModel/mpi/runtests.jl")
+end
+
+if TESTCASE ∈ ("all", "mpi", "mpi-stokes")
+  include("SurfaceStokes/mpi/runtests.jl")
 end